Add README and renamed plugins with filter prefix
