+++ plugins/config.yaml
@@ -5,7 +5,7 @@ plugins:
   - name: "PIIFilterPlugin"
     kind: "plugins.pii_filter.pii_filter.PIIFilterPlugin"
     description: "Detects and masks Personally Identifiable Information"
-    version: "1.0"
+    version: "0.1.0"
     author: "Mihai Criveti"
     hooks: ["prompt_pre_fetch", "prompt_post_fetch"]
     tags: ["security", "pii", "compliance", "filter", "gdpr", "hipaa"]
@@ -37,9 +37,9 @@ plugins:
         - "555-555-5555"
   # Self-contained Search Replace Plugin
   - name: "ReplaceBadWordsPlugin"
-    kind: "plugins.regex.search_replace.SearchReplacePlugin"
+    kind: "plugins.regex_filter.search_replace.SearchReplacePlugin"
     description: "A plugin for finding and replacing words."
-    version: "0.1"
+    version: "0.1.0"
     author: "MCP Context Forge Team"
     hooks: ["prompt_pre_fetch", "prompt_post_fetch"]
     tags: ["plugin", "transformer", "regex", "search-and-replace", "pre-post"]
@@ -57,9 +57,9 @@ plugins:
         - search: crud
           replace: yikes
   - name: "DenyListPlugin"
-    kind: "plugins.filter.deny.DenyListPlugin"
+    kind: "plugins.deny_filter.deny.DenyListPlugin"
     description: "A plugin that implements a deny list filter."
-    version: "0.1"
+    version: "0.1.0"
     author: "MCP Context Forge Team"
     hooks: ["prompt_pre_fetch"]
     tags: ["plugin", "filter", "denylist", "pre-post"]
