+++ plugins/regex_filter/plugin-manifest.yaml
@@ -1,6 +1,6 @@
 description: "Search replace plugin manifest."
 author: "MCP Context Forge Team"
-version: "0.1"
+version: "0.1.0"
 available_hooks:
   - "prompt_pre_hook"
   - "prompt_post_hook"
