+++ tests/unit/mcpgateway/plugins/fixtures/configs/valid_multiple_plugins.yaml
@@ -1,6 +1,6 @@
 plugins:
   - name: "SynonymsPlugin"
-    kind: "plugins.regex.search_replace.SearchReplacePlugin"
+    kind: "plugins.regex_filter.search_replace.SearchReplacePlugin"
     description: "A plugin for finding and replacing synonyms."
     version: "0.1"
     author: "MCP Context Forge Team"
@@ -21,7 +21,7 @@ plugins:
           replace: sullen
   # Self-contained Search Replace Plugin
   - name: "ReplaceBadWordsPlugin"
-    kind: "plugins.regex.search_replace.SearchReplacePlugin"
+    kind: "plugins.regex_filter.search_replace.SearchReplacePlugin"
     description: "A plugin for finding and replacing words."
     version: "0.1"
     author: "MCP Context Forge Team"
