+++ tests/unit/mcpgateway/plugins/fixtures/configs/valid_multiple_plugins_filter.yaml
@@ -1,7 +1,7 @@
 plugins:
   # Self-contained Deny List Plugin
   - name: "DenyListPlugin"
-    kind: "plugins.filter.deny.DenyListPlugin"
+    kind: "plugins.deny_filter.deny.DenyListPlugin"
     description: "A plugin that implements a deny list filter."
     version: "0.1"
     author: "MCP Context Forge Team"
@@ -21,7 +21,7 @@ plugins:
         - revolutionary
   # Self-contained Search Replace Plugin
   - name: "ReplaceBadWordsPlugin"
-    kind: "plugins.regex.search_replace.SearchReplacePlugin"
+    kind: "plugins.regex_filter.search_replace.SearchReplacePlugin"
     description: "A plugin for finding and replacing words."
     version: "0.1"
     author: "MCP Context Forge Team"
