+++ tests/unit/mcpgateway/plugins/fixtures/configs/valid_single_filter_plugin.yaml
@@ -1,7 +1,7 @@
 plugins:
   # Self-contained Deny List Plugin
   - name: "DenyListPlugin"
-    kind: "plugins.filter.deny.DenyListPlugin"
+    kind: "plugins.deny_filter.deny.DenyListPlugin"
     description: "A plugin that implements a deny list filter."
     version: "0.1"
     author: "MCP Context Forge Team"
