+++ tests/unit/mcpgateway/plugins/framework/test_manager.py
@@ -12,7 +12,7 @@ import pytest
 from mcpgateway.models import Message, PromptResult, Role, TextContent
 from mcpgateway.plugins.framework.manager import PluginManager
 from mcpgateway.plugins.framework.plugin_types import GlobalContext, PromptPosthookPayload, PromptPrehookPayload
-from plugins.regex.search_replace import SearchReplaceConfig
+from plugins.regex_filter.search_replace import SearchReplaceConfig
 
 
 @pytest.mark.asyncio
@@ -20,7 +20,7 @@ async def test_manager_single_transformer_prompt_plugin():
     manager = PluginManager("./tests/unit/mcpgateway/plugins/fixtures/configs/valid_single_plugin.yaml")
     await manager.initialize()
     assert manager.config.plugins[0].name == "ReplaceBadWordsPlugin"
-    assert manager.config.plugins[0].kind == "plugins.regex.search_replace.SearchReplacePlugin"
+    assert manager.config.plugins[0].kind == "plugins.regex_filter.search_replace.SearchReplacePlugin"
     assert manager.config.plugins[0].description == "A plugin for finding and replacing words."
     assert manager.config.plugins[0].version == "0.1"
     assert manager.config.plugins[0].author == "MCP Context Forge Team"
@@ -54,7 +54,7 @@ async def test_manager_multiple_transformer_preprompt_plugin():
     await manager.initialize()
     assert manager.initialized
     assert manager.config.plugins[0].name == "SynonymsPlugin"
-    assert manager.config.plugins[0].kind == "plugins.regex.search_replace.SearchReplacePlugin"
+    assert manager.config.plugins[0].kind == "plugins.regex_filter.search_replace.SearchReplacePlugin"
     assert manager.config.plugins[0].description == "A plugin for finding and replacing synonyms."
     assert manager.config.plugins[0].version == "0.1"
     assert manager.config.plugins[0].author == "MCP Context Forge Team"
@@ -66,7 +66,7 @@ async def test_manager_multiple_transformer_preprompt_plugin():
     assert srconfig.words[0].search == "happy"
     assert srconfig.words[0].replace == "gleeful"
     assert manager.config.plugins[1].name == "ReplaceBadWordsPlugin"
-    assert manager.config.plugins[1].kind == "plugins.regex.search_replace.SearchReplacePlugin"
+    assert manager.config.plugins[1].kind == "plugins.regex_filter.search_replace.SearchReplacePlugin"
     assert manager.config.plugins[1].description == "A plugin for finding and replacing words."
     assert manager.config.plugins[1].version == "0.1"
     assert manager.config.plugins[1].author == "MCP Context Forge Team"
